Split GUI classes into seperate files
